Refactor proposal: move createProposal function to sessionController
