+++ src/app.js
@@ -3,7 +3,6 @@ import cookieParser from 'cookie-parser';
 import connectDB from './config/dbConfig.js';
 import userRoutes from './routes/userRoutes.js';
 import sessionRoutes from './routes/sessionRoutes.js';
-import proposalRoutes from './routes/proposalRoutes.js';
 import adminRoutes from './routes/adminRoutes.js';
 
 const app = express();
@@ -13,7 +12,6 @@ app.use(cookieParser());
 
 app.use('/api/v1/user', userRoutes);
 app.use('/api/v1/session', sessionRoutes);
-app.use('/api/v1/proposal', proposalRoutes);
 app.use('/api/v1/admin', adminRoutes);
 
 connectDB();
